First commit of Fantasy Online 2 Java Server.
This is a basic setup of Netty using Google Protocol Buffers. Everything else is not anywhere near final.

FantasyMessages is auto-created.

FantasyServerPipelineFactory and the google protobuf code are the only things are that are probably done.

FantasyServerHandler is the entry way for all messages once a user is connected to the Netty code via TCP sockets. It also has the beginnings of a timed update thread, the start of an entity system, and a couple other things that are likely to be ripped out as I'm just testing ideas.